requirements.txt
langchain>=0.1.0
langchain-community>=0.0.6
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
websocket-client>=1.6.4
torch>=2.1.0
transformers>=4.35.0
